Added selectable friends feature!
