+++ client/src/components/FriendEntry.jsx
@@ -3,28 +3,9 @@
 import React, {
   useState, useEffect, createContext, useContext,
 } from 'react';
-import { AppContext } from './App.jsx';
 
 const FriendEntry = ({ friendName }) => {
   const [active, setActive] = useState(false);
-  const { userEmail, selectedFriends, setSelectedFriends } = useContext(AppContext);
-
-  const userStatusChange = () => {
-    const allSelectedFriends = [];
-    const activeFriend = document.getElementsByClassName('active');
-    if (activeFriend.length) {
-      for (let i = 0; i < activeFriend.length; i++) {
-        const eachFriend = activeFriend[i];
-        allSelectedFriends.push(eachFriend.value);
-      }
-      setSelectedFriends(allSelectedFriends);
-      console.log('selected friends state: ', selectedFriends);
-    }
-  };
-
-  useEffect(() => {
-    userStatusChange();
-  }, [active]); // invoke the userStatus change func when a user is selected - render delay 1 turn?
 
   return (
     <div className="eachFriend">
@@ -32,7 +13,7 @@ const FriendEntry = ({ friendName }) => {
       <input
         type="checkbox"
         onChange={(e) => {
-          console.log(`${friendName} 's box was selected`);
+          // console.log(`${friendName} 's box was selected`);
           if (active) {
             setActive(false);
           } else {
